CR-digest/requirements.txt
google-cloud-bigquery
pandas
pandas-gbq
google-generativeai
telethon
python-dotenv
